--- server/src/modules/auth/authActions.ts
@@ -1,26 +1,0 @@
-import argon2 from "argon2";
-import type { RequestHandler } from "express";
-
-const hashingOptions = {
-  type: argon2.argon2id,
-  memoryCost: 19 * 2 ** 10,
-  timeCost: 2,
-  parallelism: 1,
-};
-
-const hashPassword: RequestHandler = async (req, res, next) => {
-  try {
-    const { password } = req.body;
-    if (!password) {
-      res.status(400).json({ error: "password invalid" });
-    }
-    const hashedPassword = await argon2.hash(password, hashingOptions);
-    req.body.hashed_password = hashedPassword;
-    req.body.password = undefined;
-    next();
-  } catch (err) {
-    next(err);
-  }
-};
-
-export default { hashPassword };
